refactor: separate the LLM module from the Polkadot module
