+++ src/polkadot/index.ts
@@ -1,0 +1,5 @@
+// Polkadot Module exports
+export * from "./chain";
+export { ApiConnection } from "../common/types/typeAgent";
+export { CHAINS, XcmTransferParams } from "../common/types/xcmTypes";
+export * from "../common/types/connect";
